Module ic_consensus::consensus

This module encapsulates all components required for establishing of a distributed consensus.

Modules

payload_builder
Payload creation/validation subcomponent

pool_reader
Wrapper to read the consensus pool

utils
Consensus utility functions

Structs

ConsensusGossipImpl
Implement Consensus Gossip interface.

ConsensusImpl
ConsensusImpl holds all consensus subcomponents, and implements the Consensus trait by calling each subcomponent in round-robin manner.

Membership
Allow a node to determine what its roles are for the current round, e.g. what its rank is and what committees it belongs to.

Constants

HALT_AFTER_REGISTRY_UNREACHABLE
The maximum duration that we allow the registry to be outdated. If a subnet has not managed to get a certified statement from the registry for longer than this, the subnet should halt.

Traits

ConsensusCrypto
A trait that encompass all crypto signing/verification interface required by consensus. Anything that implements the Crypto trait automatically implements this trait.

Functions

generate_responses_to_subnet_calls
This function creates responses to the message routing with computed DKG key material for remote subnets.

setup
Setup consensus component, and return two objects satisfying the Consensus and ConsensusGossip interfaces respectively.