Trait ic_consensus::consensus::payload_builder::PayloadBuilder

pub trait PayloadBuilder: Send + Sync {
    fn get_payload(
        &self, 
        height: Height, 
        ingress_pool: &dyn IngressPoolSelect, 
        past_payloads: &[(Height, Time, Payload)], 
        context: &ValidationContext
    ) -> Result<BatchPayload, XNetPayloadError>;

    fn validate_payload(
        &self, 
        payload: &Payload, 
        past_payloads: &[(Height, Time, Payload)], 
        context: &ValidationContext
    ) -> ValidationResult<PayloadValidationError>;
}

The PayloadBuilder is responsible for creating and validating payload that is included in consensus blocks.

Required methods

fn get_payload(
&self,
height: Height,
ingress_pool: &dyn IngressPoolSelect,
past_payloads: &[(Height, Time, Payload)],
context: &ValidationContext
) -> Result<BatchPayload, XNetPayloadError>

Produces a payload that is valid given past_payloads and context.

past_payloads contains the Payloads from all blocks above the certified height provided in context, in descending block height order.

It returns a BatchPayload if the payload building is success, or an error if it fails (which at the moment is only a transient error from building XNet payload, but will be extended in the future).

fn validate_payload(
&self,
payload: &Payload,
past_payloads: &[(Height, Time, Payload)],
context: &ValidationContext
) -> ValidationResult<PayloadValidationError>

Checks whether the provided payload is valid given past_payloads and context.

past_payloads contains the Payloads from all blocks above the certified height provided in context, in descending block height order.

Loading content...