Struct ic_consensus::consensus::pool_reader::PoolReader

pub struct PoolReader<'a> { /* fields omitted */ }

A struct and corresponding impl with helper methods to obtain particular artifacts/messages from the artifact pool.

An important invariant is that the validated pool always has at least one valid random beacon and one valid block that is considered finalized.

Implementations

impl<'a> PoolReader<'a>

pub fn new(pool: &'a dyn ConsensusPool) -> Self

Create a PoolReader for a ConsensusPool.

pub fn as_cache(&self) -> &dyn ConsensusPoolCache

Return a ConsensusPoolCache reference.

pub fn registry_version(&self, height: Height) -> Option<RegistryVersion>

Return the registry version to be used for the given height. Note that this can only look up for height that is greater than or equal to the latest catch-up package height, otherwise an error is returned.

pub fn follow_to_height(&self, block: Block, height: Height) -> Option<Block>

Follow block's ancestors until a block with height is found.

pub fn chain_iterator(
&self,
block: Block
) -> Box<dyn Iterator<Item = Block> + 'a>

Find ancestor blocks of block, and return an iterator that starts from block and ends when a parent is not found (e.g. genesis).

pub fn get_range(
&'a self,
block: Block,
min: Height,
max: Height
) -> Box<dyn Iterator<Item = Block> + 'a>

Get the range of ancestor blocks of block specified (inclusively) by min and max. This assumes the correctness of the state of the pool.

pub fn get_payloads_from_height(
&self,
start: Height,
target: Block
) -> Vec<(Height, Time, Payload)>

Return a Vec of all of the Payload between the provided start height and the target block. The result is empty if target block height < start.

Note that the returned payloads are in reverse order (with decreasing heights).

pub fn get_parent(&self, child: &Block) -> Option<Block>

Returns the parent of the given block if there exist one.

pub fn get_block(
&self,
hash: &CryptoHashOf<Block>,
h: Height
) -> Result<Block, OnlyError>

Return a valid block with the matching hash and height if it exists.

pub fn get_notarized_block(
&self,
hash: &CryptoHashOf<Block>,
h: Height
) -> Result<Block, OnlyError>

Return a valid notarized block with the matching hash and height if it exists.

pub fn get_finalized_block(&self, h: Height) -> Option<Block>

Return the finalized block of a given height which is either the genesis (or CatchUpPackage) block, or the parent of another finalized block, or one with a valid finalization signature. Or return None if not found.

pub fn get_notarized_blocks(
&'a self,
h: Height
) -> Box<dyn Iterator<Item = Block> + 'a>

Return all valid notarized blocks of a given height.

pub fn get_valid_blocks(
&self,
h: Height
) -> Box<dyn Iterator<Item = Block> + 'a>

Return all valid blocks at a given height.

pub fn get_random_beacon_height(&self) -> Height

Get the max height of all valid random beacons.

pub fn get_notarized_height(&self) -> Height

Get max height of valid notarized blocks. Note that this is different than the max height of valid notarization signatures, because the notarization signature may not exist at the height of CatchUpPackage or genesis.

pub fn get_finalized_height(&self) -> Height

Get max height of valid finalized blocks. Note that this is different than the max height of valid finalization signatures, because the finalization signature may not exist at the height of CatchUpPackage or genesis.

pub fn get_finalized_tip(&self) -> Block

Get the finalized block with greatest height.

pub fn get_highest_catch_up_package(&self) -> CatchUpPackage

Get the CatchUpPackage with greatest height.

pub fn get_highest_summary_block(&self) -> Block

Get the DKG summary block with greatest height.

pub fn get_catch_up_height(&self) -> Height

Get the greatest height of CatchUpPackage.

pub fn get_random_beacon(&self, height: Height) -> Option<RandomBeacon>

Get a valid random beacon at the given height if it exists.

pub fn get_random_beacon_tip(&self) -> RandomBeacon

Get the random beacon with greatest height.

pub fn get_round_start_time(&self, height: Height) -> Option<Time>

Get the round start time of a given height, which is the max timestamp of first notarization and random beacon of the previous height. Return None if a timestamp is not found.

pub fn get_random_beacon_shares(
&self,
h: Height
) -> Box<dyn Iterator<Item = RandomBeaconShare>>

Get all valid random beacon shares at the given height.

pub fn get_notarization_shares(
&self,
h: Height
) -> Box<dyn Iterator<Item = NotarizationShare>>

Get all valid notarization shares at the given height.

pub fn get_finalization_shares(
&self,
from: Height,
to: Height
) -> Box<dyn Iterator<Item = FinalizationShare>>

Get all valid finalization shares in the given height range, inclusive.

pub fn get_random_tape(&self, h: Height) -> Option<RandomTape>

Get the valid random tape at the given height if it exists.

pub fn get_random_tape_shares(
&self,
from: Height,
to: Height
) -> Box<dyn Iterator<Item = RandomTapeShare>>

Get all valid random tape shares at the given height.

pub fn get_catch_up_package_shares(
&self,
h: Height
) -> Box<dyn Iterator<Item = CatchUpPackageShare>>

Get all valid CatchUpPackageShares at the given height.

pub fn pool(&self) -> &dyn ConsensusPool

Get the underlying pool.

pub fn dkg_summary_block(&self, block: &Block) -> Option<Block>

Returns the DKG summary block for the given valid block (which means it extends a notarized chain). Returns None if the DKG summary block cannot be found.

pub fn get_dkg_payloads(&self) -> BTreeMap<NodeId, NiDkgDealing>

Returns the set of DKG messages (indexed by the dealer Id) from the finalized tip to the highest summary block.

pub fn get_replica_version_from_highest_catch_up_package(
&self,
registry_client: &dyn RegistryClient,
replica_config: &ReplicaConfig,
log: &ReplicaLogger
) -> Option<ReplicaVersion>

Return the replica version as recorded in the block DKG summary of the highest catch up package.

Return None in case of registry lookup failure.