Function ic_consensus::consensus::utils::aggregate

pub fn aggregate<Message: Eq + Ord + Clone + Debug + HasHeight + HasCommittee, CryptoMessage, Signature, KeySelector: Copy, CommitteeSignature, Shares: Iterator<Item = Signed<Message, Signature>>>(
    log: &ReplicaLogger, 
    membership: &Membership, 
    crypto: &dyn Aggregate<CryptoMessage, Signature, KeySelector, CommitteeSignature>, 
    selector: Box<dyn Fn(&Message) -> Option<KeySelector>>, 
    artifact_shares: Shares
) -> Vec<Signed<Message, CommitteeSignature>>

Aggregate shares into complete artifacts

Consensus receives many artifact signature shares during its lifetime. aggregate attempts to aggregate these signature shares into complete signatures for the associated content.

For example, aggregate can take a Vec<&RandomBeaconShare> and output a Vec<&RandomBeacon>

aggregate behaves as follows:

groups the shares by content, producing a map from content to a vector of signature shares, each of which signs the associated content

for each grouped (content, shares) pair, lookup the threshold and determine if there are enough shares to construct a full signature for the content

if so, attempt to construct the full signature from the shares

if a full signature can be constructed, construct an artifact with the given content and full signature

return all successfully constructed artifacts

Arguments

artifact_shares - A vector of artifact shares, e.g. Vec<&RandomBeaconShare>