Function ic_consensus::consensus::utils::get_adjusted_notary_delay

pub fn get_adjusted_notary_delay(
    membership: &Membership, 
    pool: &PoolReader, 
    state_manager: &dyn StateManager<State = ReplicatedState>, 
    log: &ReplicaLogger, 
    height: Height, 
    rank: Rank
) -> Option<Duration>

Calculate the required delay for notary based on the rank of block to notarize, adjusted by a multiplier depending the gap between finalized and notarized heights, and adjusted by how far the certified height lags behind the finalized height.