Function ic_consensus::consensus::utils::is_time_to_make_block

pub fn is_time_to_make_block(
    log: &ReplicaLogger, 
    registry_client: &dyn RegistryClient, 
    subnet_id: SubnetId, 
    pool: &PoolReader, 
    height: Height, 
    rank: Rank, 
    time_source: &dyn TimeSource
) -> bool

Return true if the time since round start is greater than the required block maker delay for the given rank.