Function ic_consensus::consensus::utils::registry_version_at_height

pub fn registry_version_at_height(
    reader: &dyn ConsensusPoolCache, 
    height: Height
) -> Option<RegistryVersion>

Return the registry version to be used for the given height. Note that this can only look up for height that is greater than or equal to the latest catch-up package height, otherwise an error is returned.