Module ic_consensus::dkg

This module defines an embedding of the dkg algorithm provided by the crypto component into the consensus algorithm that is implemented within this crate.

Structs

DkgGossipImpl
DkgGossipImpl is a placeholder for gossip related DKG interfaces.

DkgImpl
DkgImpl is responsible for holding DKG dependencies and for responding to changes in the consensus and DKG pool.

Enums

PermanentError
Transient Dkg message validation errors.

TransientError
Permanent Dkg message validation errors.

Functions

create_payload
Creates the DKG payload for a new block proposal with the given parent. If the new height corresponds to a new DKG start interval, creates a summary, otherwise it creates a payload containing new dealing for the current interval.

get_configs_for_local_transcripts
Creates DKG configs for the local subnet for the next DKG intervals.

make_genesis_summary
Generates the summary for the genesis block.

make_registry_cup
Construcs a genesis/recovery CUP from the CUP contents associated with the given subnet

validate_payload
Validates the DKG payload. The parent block is expected to be a valid block.

Type Definitions

DkgMessageValidationError
Dkg errors.