Function ic_consensus::dkg::create_payload

pub fn create_payload(
    subnet_id: SubnetId, 
    registry_client: &dyn RegistryClient, 
    crypto: &dyn ConsensusCrypto, 
    pool_reader: &PoolReader, 
    dkg_pool: Arc<RwLock<dyn DkgPool>>, 
    parent: Block, 
    state_manager: &dyn StateManager<State = ReplicatedState>, 
    validation_context: &ValidationContext, 
    logger: ReplicaLogger, 
    max_dealings_per_block: usize, 
    dealing_age_threshold_ms: u64
) -> Result<Payload, TransientError>

Creates the DKG payload for a new block proposal with the given parent. If the new height corresponds to a new DKG start interval, creates a summary, otherwise it creates a payload containing new dealing for the current interval.